fix(CSI-384): manage node labels creates excessive Kubernetes API calls
